Removed yaml support from v1, tests and helper methods
