+++ config/v1/config.go
@@ -1,19 +1,15 @@
 package v1
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"reflect"
 	"strings"
-	"unicode"
 
 	"github.com/chanzuckerberg/fogg/errs"
 	"github.com/chanzuckerberg/fogg/plugins"
 	"github.com/hashicorp/go-multierror"
-	"github.com/sirupsen/logrus"
 	"gopkg.in/go-playground/validator.v9"
-	"gopkg.in/yaml.v2"
 )
 
 type TfLint struct {
@@ -151,24 +147,12 @@ func ReadConfig(b []byte) (*Config, error) {
 		Docker: true,
 	}
 
-	if IsJSON(b) {
-		e = json.Unmarshal(b, c)
-		logrus.Warn("JSON is deprecated, consider migrating to yaml")
-	} else {
-		e = yaml.Unmarshal(b, c)
-	}
+	e = json.Unmarshal(b, c)
 	if e != nil {
 		return nil, errs.WrapUser(e, "unable to parse config file")
 	}
 
 	return c, nil
-}
-
-func IsJSON(b []byte) bool {
-	jsonPrefix := []byte("{")
-	trimmed := bytes.TrimLeftFunc(b, unicode.IsSpace)
-
-	return bytes.HasPrefix(trimmed, jsonPrefix)
 }
 
 // Validate validates the config
