+++ config/v1/config_test.go
@@ -23,23 +23,24 @@ func TestComponentKindGetOrDefault(t *testing.T) {
 }
 
 func TestParseDefaults(t *testing.T) {
-	yaml :=
-		`
-defaults:
-  aws_profile_backend: czi
-  aws_profile_provider: czi
-  aws_region_backend: us-west-2
-  aws_region_provider: us-west-1
-  infra_dynamo_db_table: the-table
-  infra_s3_bucket: the-bucket
-  project: test-project
-  terraform_version: 0.11.0
-`
+	json := `
+	{
+		"defaults": {
+			"aws_region_backend": "us-west-2",
+			"aws_region_provider": "us-west-1",
+			"aws_profile_backend": "czi",
+			"aws_profile_provider": "czi",
+			"infra_s3_bucket": "the-bucket",
+			"infra_dynamo_db_table": "the-table",
+			"project": "test-project",
+			"terraform_version": "0.11.0"
+		}
+	}`
 
-	r := ioutil.NopCloser(strings.NewReader(yaml))
+	r := ioutil.NopCloser(strings.NewReader(json))
 	defer r.Close()
 
-	c, e := ReadConfig([]byte(yaml))
+	c, e := ReadConfig([]byte(json))
 	assert.NoError(t, e)
 
 	assert.Equal(t, "czi", c.Defaults.AWSProfileBackend)
@@ -93,17 +94,6 @@ func TestParse(t *testing.T) {
 	assert.NotNil(t, c.Modules)
 }
 
-func TestYamlFailure(t *testing.T) {
-	yaml := `foo`
-	r := ioutil.NopCloser(strings.NewReader(yaml))
-	defer r.Close()
-	b, e := ioutil.ReadAll(r)
-	assert.NoError(t, e)
-
-	c, e := ReadConfig(b)
-	assert.Nil(t, c)
-	assert.NotNil(t, e)
-}
 func TestJsonFailure(t *testing.T) {
 	yaml := `foo`
 	r := ioutil.NopCloser(strings.NewReader(yaml))
@@ -139,25 +129,27 @@ func TestValidation(t *testing.T) {
 }
 
 func TestExtraVarsValidation(t *testing.T) {
-	yaml := `
-defaults:
-  account_id: 123456789
-  aws_profile_backend: czi
-  aws_profile_provider: czi
-  aws_provider_version: czi
-  aws_region_backend: us-west-2
-  aws_region_provider: us-west-1
-  infra_dynamo_db_table: the-table
-  infra_s3_bucket: the-bucket
-  owner: test@test.com
-  project: test-project
-  terraform_version: 0.11.0
-`
+	json := `
+	{
+		"defaults": {
+			"aws_region_backend": "us-west-2",
+			"account_id": 123456789,
+			"aws_region_provider": "us-west-1",
+			"aws_profile_backend": "czi",
+			"aws_profile_provider": "czi",
+			"aws_provider_version": "czi",
+			"infra_s3_bucket": "the-bucket",
+			"infra_dynamo_db_table": "the-table",
+			"project": "test-project",
+			"owner": "test@test.com",
+			"terraform_version": "0.11.0"
+		}
+	}`
 
-	r := ioutil.NopCloser(strings.NewReader(yaml))
+	r := ioutil.NopCloser(strings.NewReader(json))
 	defer r.Close()
 
-	c, e := ReadConfig([]byte(yaml))
+	c, e := ReadConfig([]byte(json))
 	assert.Nil(t, e)
 
 	e = c.Validate()
